+++ packages/client-config/src/model_introspection_schema_adapter.test.ts
@@ -1,12 +1,18 @@
 import assert from 'assert';
 import { beforeEach, describe, it } from 'node:test';
-import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
 import { mockClient } from 'aws-sdk-client-mock';
 import { sdkStreamMixin } from '@aws-sdk/util-stream-node';
 import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3';
 import { Readable } from 'node:stream';
 import { ModelIntrospectionSchemaAdapter } from './model_introspection_schema_adapter.js';
+import { AmplifyClient } from '@aws-sdk/client-amplify';
+import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
 
+const stubClientProvider = {
+  getS3Client: () => new S3Client(),
+  getAmplifyClient: () => new AmplifyClient(),
+  getCloudFormationClient: () => new CloudFormationClient(),
+};
 void describe('ModelIntrospectionSchemaAdapter', () => {
   const s3Mock = mockClient(S3Client);
 
@@ -16,7 +22,7 @@ void describe('ModelIntrospectionSchemaAdapter', () => {
 
   void it('returns undefined on undefined schema uri', async () => {
     const modelIntrospectionSchema = await new ModelIntrospectionSchemaAdapter(
-      fromNodeProviderChain()
+      stubClientProvider
     ).getModelIntrospectionSchemaFromS3Uri(undefined);
     assert.deepEqual(modelIntrospectionSchema, undefined);
   });
@@ -24,7 +30,7 @@ void describe('ModelIntrospectionSchemaAdapter', () => {
   void it('throws on invalid s3 location', async () => {
     await assert.rejects(() =>
       new ModelIntrospectionSchemaAdapter(
-        fromNodeProviderChain()
+        stubClientProvider
       ).getModelIntrospectionSchemaFromS3Uri('s3://im_a_fake_bucket/i_swear')
     );
   });
@@ -51,7 +57,7 @@ void describe('ModelIntrospectionSchemaAdapter', () => {
       });
 
     const modelIntrospectionSchema = await new ModelIntrospectionSchemaAdapter(
-      fromNodeProviderChain()
+      stubClientProvider
     ).getModelIntrospectionSchemaFromS3Uri('s3://im_a_real_bucket/i_swear');
     assert.deepEqual(modelIntrospectionSchema, {
       version: 1,