+++ packages/plugin-types/src/index.ts
@@ -19,3 +19,4 @@ export * from './package_manager_controller.js';
 export * from './deep_partial.js';
 export * from './stable_backend_identifiers.js';
 export * from './resource_name_validator.js';
+export * from './aws_client_provider.js';